requirements.txt
psycopg2-binary==2.9.9
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-crum==0.7.9
idna==3.10
plyer==2.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
